requirements.txt
gradio>=4.10.0
sentence-transformers>=2.2.2
numpy>=1.23
requests>=2.31.0
pdfplumber>=0.9.0
pdf2image>=1.17.0
pytesseract>=0.3.10
Pillow>=10.0.0
pymupdf>=1.23.0
easyocr>=1.7.1
opencv-python-headless>=4.8.0
python-docx>=0.8.11
